fade_out_hatch_instructions(): Add
A new function saying how to make the 'instructions' paragraph fade out.
The `fadeOut()` function is part of the `jQuery` library we're using and
takes care of the animation, timing, etc.  Again, nothing happens yet
because all we've done is say what `fade_out_hatch_instructions` means.
So far, we haven't told it to *do* those instructions.