Add health, happiness, hungriness level displays
Add three `<p>` elements to the html, to show the Tamagotchi's current
levels of health, happiness, and hungriness.  It starts off fully
healthy and happy, and not at all hungry.

This isn't quite right, because these displays are always visible, even
at the very start when only the egg is on the screen.  The displays
should only appear when the Tamagotchi has hatched; we'll fix that next.