Temp: Make alien hungrier 2" after hatching
Just to check that this is all working, we'll set up a timer to make the
Tamagotchi get hungrier two seconds after it hatches.  We'll get rid of
this once we're happy it's all working correctly.  At this point, if you
hatch the Tamagotchi, you should then see the hungriness go up from 0 to
5 after two seconds.