time_goes_by(): Add
For now, the only thing that will happen as time goes by is that the
Tamagotchi gets hungrier.  As we develop the game further, it might also
catch a disease (lose health), or become sad (if not played with
enough).

But for now, nothing actually happens, because we haven't said that we
want `time_goes_by()` to take place; we've only defined it.