alien_is_alive(): Add
We need to know if the Tamagotchi has got so hungry that it dies of
starvation.  To check this, we define a new function.  This function is
different to the ones we've writtem so far, because it gives back a
piece of information to the caller.  This is known as 'returning' the
information.  Here, the information is whether the Tamagotchi is still
alive, and JavaScript has the idea of a *Boolean* value for this &mdash;
a value which is either *true* or *false*.  In our case, whether the
Tamagotchi is alive is worked out by asking 'is its hungriness less than
100?'.

Nothing uses this function yet, though.