<s>Add button to give bread to Tamagotchi
We'll add a way for the player to feed the Tamagotchi.  Looking ahead a
bit, we'll need to add more than one button, so we'll add a `<p>` to
hold them all, and add it below the levels displays.

Then, we'll use the HTML `<button>` element to create the actual button.

This doesn't look right yet, because it should be hidden at the start of
the game.  We'll fix that next.