Fade in action buttons with levels display
We add a second call to `fadeIn`, for the `action-buttons` element.  The
two fade-in animations happen at the same time, which is
appropriate.  (If we wanted one to fade in and then the other, we'd add
extra arguments to the first `fadeIn` call, like we did to make the
displays fade in after the hatch instructions had finished fading out.)