Make sure hungriness doesn't become negative
There's a bug!  You can make the Tamagotchi very UN-hungry by repeatedly
feeding it bread.  The hungriness can become very negative, which
doesn't make sense.  We'll fix this now.

After reducing the hungriness, we check whether we've reduced it so far
that it's become less than zero.  If so, change it to be zero.