Draw smiling mouth
To give the Tamagotchi a mouth, we create a path and then `stroke()`
it.  This just gives the outline, not a filled-in shape.  To build the
path, we move to the starting point, then draw three lines to the other
corners in the shape.  Experiment with the shape, colour, width, etc.