feed_alien_sweets(): Add
We can copy-and-paste the code from the `feed_alien_some_bread()`
function to describe what to do to feed the Tamagotchi sweets.  The only
difference is that sweets only reduce the hungriness by 10, not 20.

This is untidy, though, and makes our code more difficult to maintain
and change, so we will fix this near-repetition shortly.