`maybe_make_alien_sicker()`: Add skeleton
We will fill in the details in a minute, but first we set up the
structure of how to do something randomly.  JavaScript provides a
`Math.random()` function, which gives you a random number between 0 and
1 each time you call it.  By testing whether this random number is less
than 0.1 (i.e., 1/10), we can make something happen about one time in
ten that we call the function.