<s>Draw Tamagotchi sickly green if not very healthy
To make the game look more interesting, and also slightly easier, we'll
draw the Tamagotchi a different colour when it's sick.  The `rgb()`
notation lets you describe a colour using red, green, and blue parts in
the same way as the `#e3f` notation in CSS, except here you give three
whole numbers between 0 and 255.