Check whether Tamagotchi has reached winning age
After updating the age, we need to check if the player has won.  This
needs to fit in with the '*has the player lost?*' check, so we'll
rearrange this part of `time_goes_by()` a bit.

For testing, it's a good idea to use a smaller number than 100 (maybe
10), otherwise it takes an annoyingly long time to check it's working.

If, by coincidence, the alien reaches old age at the same time as
getting to zero health (say), we give the player the benefit of the
doubt and call it a win.  We do this by checking the age before the
`alien_is_alive()` check.