Introduction to web programming
Done some Scratch?  Got a lot of useful ideas from that?  Let's see how
to take similar ideas and create something which runs on the web.  We
will work towards a Tamagotchi-style game.

In some ways programming on the web is very similar to Scratch &mdash;
Scratch has sprites with costumes and scripts.  The web-app will have
parts that describe what the page looks like (like costumes), and parts
which say how it behaves (like scripts).