<s>Get a '2D drawing context' for the canvas
Our drawing is going to be in 2D.  You can do 3D drawing, but it's quite
a lot more complicated.  We find the canvas using the name (`id`) we gave
it, and then ask the canvas for its '2D' drawing 'context'.  It's not
hugely important right now exactly what a 'context' is, but it's what
lets us draw on the page.

The `document.getElementById()` finds the `game` element, which is the
canvas we created in the HTML earlier.

The dot (`.`) between the `canvas` and the `getContext` is a bit like
choosing the `getContext` option from a right-click menu of things you
can do with or to `canvas`.  The `()` are for making it actually do the
set of instructions (function) `getContext`.