Draw Tamagotchi 'egg'
The egg is a simple circle, but we need to know about 'paths' to draw
one.  You set up a path that an imaginary pen is going to take round the
canvas, and then, when it's done, say whether to: draw the outline (with
a 'stroke' of the pen) or 'fill' it in.  We want a coloured-in circle so
will 'fill' it; you can choose a colour, or something more
complicated (gradient or pattern) &mdash; we'll stick with a plain
colour.

To talk about position on the canvas, everything is done in terms of _x_
and _y_ coordinates &mdash; starting in the top-left corner, how far
across should you go (_x_), and then how far down (_y_).  Annoyingly,
this is different to how Scratch does it, where you start in the middle
and then go across and up.  Scratch's way is how mathematicians do
things; most computer graphics is done the way the HTML canvas does it.
This is for 'historical reasons'.