Add 'Hatch' button and instructions
We can add a button which will start the game using the HTML `<button>`
element.  But so far, nothing happens when you click on it, because we
haven't connected it to any JavaScript code.

Add styling to the button if you like by adding `#hatch {...}` to the
CSS.  The identifier for the button is `hatch`.