Say we want to draw the alien when button clicked
We declare that when the button (which we describe by giving its id) is
clicked, we want the function (set of instructions) `draw_alien` to
happen.  But we haven't yet said what `draw_alien` means, so it'll look
like nothing happens when we click the button.  If you bring up the
'developer tools' for your browser, you can see an error:  `Uncaught
ReferenceError: draw_alien is not defined` which means you're trying to
refer to something but you haven't defined it.  We'll do that next.

The current situation is like an imaginary situation in Scratch where
you've got a script which says 'when this sprite clicked,
jump-up-and-down', but you haven't defined the custom 'jump-up-and-down'
block.  (This situation is imaginary because Scratch, sensibly, won't
let you do this.)