</s>Draw the egg and make it hatch
The first step is to show the Tamagotchi's egg when the player loads our
web app.  Then, when the player clicks a button, the egg will hatch into
the actual Tamagotchi.

To do this, we'll need to add the button to the `HTML`, and then write
some JavaScript code to do the drawing, and connect them together.