</s>Define function to draw a circle
Take a step back and notice that we're drawing circles three times, and
repeating the same slightly fiddly set of instructions each time.  When
you notice this, it's a good idea to collect those instructions together
and give them a name, i.e., define a new function.  We've done this
before, but this function has something new compared to the ones we've
defined before &mdash; it's not enough to just say 'draw a circle', you
need to know *where* and *how big*. The function has extra information it
needs before it can do its job.  The 'caller' has to supply those extra
pieces of information.

That's why our previous functions have `()` &mdash; there's nothing
between the `()` because there is no extra information required.

Once we have this handy way of saying 'draw a circle here and make it
this big', we use it to make the rest of our code easier to read.

Note that having made these changes, the app behaves the same, but it is
easier to read and understand.  This is known as *code refactoring*.