</s>Make Tamagotchi get hungrier (basic idea)
To make a start on the game mechanics, we'll set up the basic idea of
keeping track of the Tamagotchi's changing levels of hunger.  We add a
variable in the JavaScript, to keep track of how hungry it is, and write
a function to update both it and the display.  Temporarily, to make sure
everything is working, we make the Tamagotchi get 5 steps hungrier after
two seconds.