</s>Allow choice of feeding bread or sweets
So far the player can only feed bread to the Tamagotchi.  We will give
the player the choice of feeding sweets instead, by adding another
button and writing a new function.  We can then notice that the
bread-feeding function and the sweets-feeding function are very similar,
so we can *refactor* by writing a general function to feed either food.