`temporarily_disable()`: Add
We are going to want to apply the same 'disable then re-enable'
behaviour to the 'feed sweets' button, so make this into a function.
One wrinkle is that we can't (directly) define a general equivalent of
the `enable_bread_button()` function, so instead we'll use an *anonymous
function* in the `setTimeout()` call.

Our new function needs to know which button to work with, so this is the
argument &mdash; callers will pass something like `'#feed-bread'`.