</s>Stop the player hammering the feeding buttons
At the moment, the player can just keep clicking the *Bread*
and *Sweets* buttons as quickly as they like, which makes the game too
easy.  We will make it so the feeding buttons are disabled for a short
while after being clicked.

Because we want this behaviour for both buttons, we'll write a function
for it, and then use it for both the buttons.