Make sweets reduce the Tamagotchi's health
As part of what happens when the player feeds the Tamagotchi some
sweets, we want to reduce its health.  In the same way as we handled the
hungriness, there are two parts:

* Update the `health` variable.
* Update the displays in the web page with the new value.

There is a `TODO` to remind us to do something: in this case, we want to
check whether the Tamagotchi is so unhealthy that it dies.  We'll fill
this in later.