<s>game_is_running(): Add
Looking ahead a bit, the player will be able to win the game by
nurturing their Tamagotchi to old age.  We therefore need a function
which tells us whether the game is still running.  (The alternative
being that it has stopped, either through the player losing or through
the player winning.)