Do not update game state if game is not running
Once the game has stopped, clicking on buttons shouldn't do anything,
and the 'time goes by' function shouldn't do anything either.  To
achieve this, we check whether `game_is_running()` as the first thing we
do on entering these functions:

* `feed_alien_some_bread()`
* `feed_alien_sweets()`
* `time_goes_by()`

If the game isn't running, we `return` straight away, which means we
don't do any more of the instructions in that function.