</s>Random illnesses
Occasionally, the Tamagotchi gets sick just through bad luck.  We
already have a function to make things happen every 'tick' of the game
clock, so we can add a 'randomly make sicker' feature to it.  The
randomness comes from JavaScript's `Math.random()` function, and to keep
the code easy to read, we'll put this into a separate new function.