`make_alien_sicker()`: Add
Copying the common behaviour between `maybe_make_alien_sicker()` and
`feed_alien_sweets()`, we see that the difference is in how much health
the Tamagotchi loses.  We therefore make this amount be an argument to
our new function.

One other difference is that we never checked for `health < 0` in
`feed_alien_sweets()`, which we got away with because `health` was
always a multiple of 5.  It's no harm to check, though, and in future we
might not always have `health` being a multiple of 5.