</s>Create one function to make Tamagotchi sicker
There is a large overlap between the code which reduces the Tamagotchi's
health when the player feeds it a sweet, and the code which reduces its
health because of random sickness.  To make the code easier to work
with, we can *refactor* by pulling that common logic out into a new
function.