</s>Make player win game if Tamagotchi reaches old age
So far the game only has ways to lose, and not to win.  We'll make it so
that the player wins the game if the Tamagotchi reaches the ripe old age
of 100.  This will involve adding an 'age' display, and keeping track of
the age via a JavaScript variable.  The age needs to increase as time
goes by, and we need to check whether the Tamagotchi becomes old enough
to tell the player they've won.