Use alien_still_growing() in time_goes_by()
We don't want to have two places in the code where we decide whether the
alien has reached old age.  This kind of thing inevitably leads to
trouble when you change things in one place and not the other.  To avoid
this, use the `alien_still_growing()` function to decide whether the
player has won &mdash; if it's *not* still growing, the player has won.