update_alien_happiness(): Add
Being bored makes the Tamagotchi less happy.  If it's not bored, it has
a generally positive outlook on life, so its happiness goes up.  We need
a function to check whether the alien is bored, and then move its
happiness up or down accordingly.  This function has to do quite a lot.

The main bit is the `((age - last_game_age) >= 3)` test.  This
subtraction works out how long ago the alien last had a game, and if
that was three or more hours ago, it's bored.  Then we adjust the
`happiness` variable either down (if bored) or up (if not), and update
the display.  We also check that the `happiness` doesn't become less
than zero or more than 100.