Become un-bored immediately when playing a game
When the player clicks the *Game* button, the Tamagotchi doesn't
immediately become non-bored, only on the next game tick.  This isn't a
good game experience, so update the happiness (which re-draws the
Tamagotchi) as part of responding to the *Game* click.  This also gives
an immediate `happiness` boost, which is probably appropriate.