Make Tamagotchi look unhappy if too unhappy
We can turn the smile upside down into a sad face if the Tamagotchi is
too unhappy.  We check this first, so if the Tamagotchi is both very
unhappy and bored, it looks sad.

We're already calling `draw_alien()` when the `happiness` variable might
change, because this is tied in with the `bored` variable.