</s>Make Tamagotchi's appearance show its state
The game will look more appealing, and also the player will have a
better chance of keeping an eye on what's going on, if we make the
Tamagotchi look different according to how hungry, happy, and healthy it
is.  We change `draw_alien()` quite a bit, to show:

* A sad, bored, or happy mouth.
* A sickly green colour if it's ill.
* A hole in its belly as it gets hungrier.

The code needs updating to call `draw_alien()` at various places, to
make sure the appearance always matches the variables and levels
display.