`update_display()`: Add
Collect the various updates to the different bits of the webpage
together into one function.  When we call this function, many parts will
not have changed, so we will be doing more work than required.  But
computers are fast these days, and making the code simpler to understand
is usually more important than having it run as quickly as possible.