Call update_display() where required
Every part of the code which changes the game's state now needs to call
`update_display()`.  These parts are the 'time goes by' code, and any
user-driven behaviour (button clicks):

* `feed_alien_some_bread()`
* `feed_alien_sweets()`
* `give_alien_medicine()`
* `play_game()`
* `time_goes_by()`

We have temporarily broken the timer code, because it was tied up with
the display update for win/lose.  We'll fix that shortly.