</s>Simplify display-update logic
The current logic for making sure the display is up-to-date is a bit
fragile and scattered.  Every time we change a variable, we update that
part of the display.  Simpler would be to update the whole display after
anything might have changed.

We collect all the display-update code into one new function,
`update_display()`, and call it whenever things might have changed.