Add variable to remember where wrappers are
JavaScript variables can be *lists*, properly known in JavaScript
as *arrays*, which means they can hold more than one object.  Scratch
has the same idea.  At the start of the game, there are no wrappers, so
we have a list with nothing in it.  This is known as an *empty list*.