Draw the wrappers
We have a `draw_alien()` function, which we'll give the job of drawing
the wrappers, even though they're not really part of the alien.

We want to draw each wrapper, and one way to do this in JavaScript is to
use the `forEach()` method on the list of things.

In this code, the `w` variable is set (by JavaScript) to be equal to
each of the things in `sweet_wrappers` in turn, and the `fillRect()` is
done for each one.  Remember that each thing in the list is itself a
list, with the *x* and *y* coordinates as its two entries.  You get
things out of this list with `w[0]` to get the first, and `w[1]` for the
second.  JavaScript counts list entries starting from zero.  Here I've
just drawn a red rectangular sweet wrapper but you can do something more
complicated if you like.

If you try the game now, you should see the two test sweet-wrappers we
set up.