Make up random position for a new wrapper
When the Tamagotchi gets a sweet, we'll need to make up a random
position for it to drop the wrapper.  We'll use the same `Math.random()`
function as we've done before (when we were randomly making it sick).
But this time we need to turn our random number between 0 and 1 into a
sensible position on the canvas, which we do by multiplying by the range
of possible values, then adding the smallest desired value.

Here's I've made it so the wrappers will only appear in the bottom half
of the screen, by making `wrapper_y` be at least 240.

So far this function makes up the `x` and `y` coordinates of the
wrapper's position, but doesn't yet add the new wrapper to the list.