Wrappers make the Tamagotchi unhappy
The messier its home is, the quicker it becomes unhappy.  We'll subtract
happiness points equal to the number of wrappers littering the place.
To do this, we use the `length` property of `sweet_wrappers`, which
tells us how many things are in the array.

As usual, we have to make sure that we don't leave the `happiness`
negative.