`sweep_up_wrappers()`: Add
When the player clicks the button, we're going to want to remove all the
sweet wrappers.  The easiest way to do this is just to make
`sweet_wrappers` be a fresh empty array.

We also need to temporarily disable the *Sweep* button, and also update
the display, to make sure the clean room is immediately shown.