</s>Add 'sweet wrappers' features
When the Tamagotchi eats sweets, it drops the wrappers on the floor of
its room.  The untidiness makes it unhappy, and it gets unhappier more
quickly the more wrappers are littering its room.  To allow the player
to fix this, we add a *Sweep* button, which clears up all the wrappers.